third_party/blink/web_tests/external/wpt/clear-site-data/set-cookie-before-clear-all.https-expected.txt

This is a testharness.js-based test.
FAIL Setting a cookie before Clear-Site-Data for all shouldn't allow the cookie to take. assert_equals: expected "" but got "testSetWithClear=true"
Harness: the test ran to completion.
